feature(typescript): detect typescript dependencies that only exist before compile time (#30)
* feature(typescript): use the typescript AST to infer some of the dependencies:
Some typescript features and dependencies get lost when the typescript compiler translates to javascript. By using the typescript AST instead of the javascript-AST-after-translation (e.g. when you depend only on a type from a module, if you depend on a module that isn't used (yet), have triple-slash directives, or use the import-from-export-equals construct) , we'll be able to catch these as well. Will solve issue #28

* refactor(extract): pull out the 'parse' step from extract.js and ...
  * ... rename all modules that extract dependencies from an AST from extract-<modulesystem> to etract-<modulesystem>-deps as that better conveys their purpose
  * ... put the extract-yadda-deps in a separate folder
  * ... separate parsing and extracting in the (new) typescript dependency extraction
* style(depcruise): updates some of our own dependency-cruiser rules
* test(extract): add a test for when dependencies aren't used and they still should show up in the output
* doc(parse): document the parser wrappers
* doc(samples): updates the tslint sample