feature(validate): set up the validation engine to correctly process orphans.
That seems like an awfull lot of changes for the set up of a rule. The reason:
the orphan rule is a _module only_ rule, whereas rules hitherto were rules
on _dependencies only_

This commit also contains a sneaky refactoring: on the top level of the
(json, bare api) output the array of modules is now called `modules`
in stead of `dependencies` - which is more correct and makes a lot of
the code instantly easier to grasp.

For those using the api or the json response this will count as a
BREAKING CHANGE. If you don't, you probably wouldn't care less.