feature(main): only load the ts paths plugin when the tsconfig contains a baseUrl
This should prevent some confusion - the plugin will complain at runtime it doesn't want to load the tsconfig if there's no baseUrl. This runtime message looks like it's coming from dependency-cruiser, while it isn't. Actually everything is fine, just the paths resolution isn't done with the paths plugin (which is perfectly fine, as there was no baseUrl in the tsconfig in the first place)