doc: re-generate graphs
the graph generation kicked off with the latest version bump was run on a folder with a node_modules in an interesting state (complaints about files in .staging missing???) - maybe because of yarn was run on it in between. As a consequence the graphs had all kinds of complaints in them of unreachability and missing shizzle. This run was done on a fresh install (tm) and graphs look as they should