perf(resolve): only load tsconfig-paths-plugin when there's at least one defined path (#419)
tsconfig-paths-plugin has a measurable impact even when there's no paths in the tsconfig: ~10ms for 15 resolutions or 0.7ms per resolve operation. As there's typically a few hundred of these going on even on small projects this will save a few 100's of ms on projects that don't use paths in their tsconfigs.