bugfix(main): relax the check to load tsconfig-paths-plugin (#427)
* bugfix(main): relax the check to load tsconfig-paths-plugin

Dependency-cruiser used to check for a non-empty 'paths' object in the tsconfig (PR #419, commit 60eea5a, as that's the primary goal of the ts-configs-plugin) before enabling the ts-config-paths plugin. As it turns out tsconfig-paths also resolves fixed/path/jadda.js to ${baseDir}/fixed/path/jadda.js if the 'fixed' module or alias doesn't exist otherwise. Also at least 2 (but likely a lot more) consumers rely on that behaviour.