bugfix(extract): makes baseDir option usable from API and config (#438)
* fix bug where baseDir is ignored

When calling dependency cruiser programmatically my understanding is specifying `baseDir` in `pOptions` should be equivalent to running `depcruise` shell command in that folder. `baseDir` defaults to cwd so that is how the functionality seems to me.

But `gatherScannableFilesFromDirectory` breaks that contract because it calls `readdirSync(pDirectoryName)` which *implicitly* relies on cwd when passed a relative path. If we instead call `readdirSync(path.join(pOptions.baseDir, pDirectoryName))` then `pDirectoryName` gets resolved according to `baseDir`, which by default will be the same behavior as before but now lets the caller override it via options.

The same problem exists on L38 where stats are called.

With this fix in place I can now call this:
```
futureCruise(
  ['.'],
  { outputType: 'dot', baseDir: dirOfMyChoosing },
)
```
And see the same output as if I run:
```bash
dirOfMyChoosing $ depcruise --output-type dot .
```

* chore: adds 'baseDir' to the type definitions & json schema

* doc: document the baseDir option

* test: adds tests for the baseDir option

Co-authored-by: sverweij <sverweij@users.noreply.github.com>