bugfix(test): makes the io unit test run more reliably
before we created and removed a file before and after the tests. On the ci this ran into unstable behaviour: test runs stopped with ENOENTs on the file, either because it hadn't been created yet or because it was already removed. It puzzles me why we'd get a race condition like this in synchronous code (something in mocha? something I don't understand about node?) but now it's not possible to get it anymore because the fixture is part of the repo now. Additional benefit: it makes the test code more simple