chore: make double-sure generated meta.js is prettier'ed
we noticed src/meta.js sometimes was generated with a line-end at the end of the file, and sometimes without one. Interesting because there's only one way to generate it (a script that emits stuff to stdout). Maybe yarn/ npm do the redirect thing different from the shell? Anyway - an extra prettier --write pass in the Makefile seems to make it consistent again ...