Suggest `typings` as a main field whenever `types` is suggested
The FAQs mention that if types from a package cannot be resolved then
the user should configure `mainFields` to include the keys that the
package is using in its `package.json` file.

The generated config and documentation suggest using `main` and `types`
as `mainFields`. However, TypeScript documentation notes that:

> ... the "typings" field is synonymous with types, and could be used as well.

https://www.typescriptlang.org/docs/handbook/declaration-files/publishing.html

Rather than make users of this package search through documentation and
`node_modules` to discover this issue, this commit changes
`dependency-cruiser` to support (or at least, suggest) both of these
TypeScript defaults.