Suggest `typings` as a `mainFields` wherever `types` is suggested (#831)
First off, I just want to say thanks very much for producing and
maintaining this tool. I've been amazed at how fast/powerful it is, and
the thorough documentation in config and READMEs has been super helpful!
Thank you! ❤️ 🙇

## Description

The FAQs mention that if types from a package cannot be resolved then
the user should configure `mainFields` to include the keys that the
package is using in its `package.json` file.

The generated config and documentation suggest using `main` and `types`
as `mainFields`. However, [TypeScript
documentation](https://www.typescriptlang.org/docs/handbook/declaration-files/publishing.html)
notes that:

> ... the "typings" field is synonymous with types, and could be used as
well.


https://www.typescriptlang.org/docs/handbook/declaration-files/publishing.html

Rather than make users of this package search through documentation and
`node_modules` to discover this issue, this commit changes
`dependency-cruiser` to support (or at least, suggest) both of these
TypeScript defaults.

## Motivation and Context

One of our dependencies changed how they export types. We spent several
confusing hours trying to track down the issue. It was especially
confusing because we'd already configured dependency-cruiser to use
`["main", "types"]` for `mainFields`, as the documentation suggests.

We were not aware that TypeScript also supported `typings` as a synonym
for `types`. It would've saved us a lot of head scratching if
`dependency-cruiser` had suggested both these values from the beginning.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

- [x] green ci

## Screenshots

<!-- Only if appropriate - feel free to delete this section if it's not
applicable -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

Marking as a documentation change as it only affects documentation/the
initial config that is generated by the tool.

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [X] :balance_scale:
- The contribution will be subject to [The MIT
license](https://github.com/sverweij/dependency-cruiser/blob/main/LICENSE),
and I'm OK with that.
  - The contribution is my own original work.
- I am ok with the stuff in
[**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/main/.github/CONTRIBUTING.md).

---------

Co-authored-by: sverweij <sverweij@users.noreply.github.com>