feat(cache): adds option to compress the cache file (#860)
## Description

- adds an option to compress the cache file
- 🏕️ some updates in the cache typing

## Motivation and Context

The cache file can become quite large. The cache file is also _very_
compressible (both gzip and brotli reduce the file size to 10% of the
original _or less_). My original hunch was that as less i/o is needed
this might constitute a net performance gain for large enough code
bases. I haven't been able to substantiate that with evidence, though.

However, I decided to still let this in because:
- when you switch compression on, it (only) adds a few milliseconds to
the total execution time
- the resulting file is so much smaller, that the trade-off of the few
extra milliseconds might be worth it for other reasons

## How Has This Been Tested?

- [x] green ci
- [x] additional automated tests

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)