feat(cache): adds cache format version & checks against it (#891)
## Description

- adds a cache format version attribute
- adds a check to skip using the cache when the version found in the
cache is too low
- makes the typescript typings in some of the cache related modules a
bit more readable (and valid).

## Motivation and Context

There are some changes in the dependency-cruiser output format that are
incompatible between version 15 and 16. For use without a cache (or
_known violations_, but see #890 for that) this presents no issues. The
cache can still contain the old format, though, which might result in
slightly unexpected output or disturbing looking errors. With this
feature dependency-cruiser will ignore the contents of the cache when
the format version of that cache is too low.

## How Has This Been Tested?

- [x] green ci
- [x] additional non-regression tests

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)