fix(types): enable IDE autocompletion for OutputType candidates (#1005)
## Description

Updated the definition of `OutputType` so that IDEs can now provide
autocompletion for its literal candidates.

## Motivation and Context

Although the common candidates for ‎`OutputType` are defined as
literals, autocompletion in IDEs was not available because they were
unioned with ‎`string`.
However, for a better developer experience, it is preferable for these
literal candidates to be available as autocomplete suggestions.

## How Has This Been Tested?

- [x] green ci

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

Although no issue has been opened for this change, I have selected “Bug
fix” because it does not affect runtime behavior in any way.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
- The contribution will be subject to [The MIT
license](https://github.com/sverweij/dependency-cruiser/blob/main/LICENSE),
and I'm OK with that.
  - The contribution is my own original work.
- I am ok with the stuff in
[**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/main/.github/CONTRIBUTING.md).