refactor(enrich/derive/circular):  moves creation of indexed module graph into deriveCycles (#1034)
## Description

- refactor(enrich/derive/circular): moves creation of indexed module
graph into deriveCycles

## Motivation and Context

- because it's an "implementation detail" and
- the indexed module graph is currently used _only_ in deriveCycles, so
it's not even a performance optimization; on the contrary...
- ... now that indexing step for cycles only takes place when there's
cycles to detect

Performance-wise this might save a few centiseconds on very large repos
that donot contain rules with cycle-detection. I don't expect to be a
lot of those to be out there, so the main goal of this PR is code
cleanliness/ separation of concerns.

## How Has This Been Tested?

- [x] green ci

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)